Refactor a separate global Table component.
