+++ frontend/src/pages/cache/Cache.tsx
@@ -10,19 +10,19 @@ import {
     ContainerHeader,
     ContainerBody,
 } from './../../components/Containers'
-import { CacheTable } from './../../components/Tables'
+import CacheTable from './components/Table'
 
 // local components
 import HeaderForm from './components/HeaderForm'
 
 // interfaces
-import {
-    CacheInterface,
-    CacheResponseInterface,
-} from './../../interfaces/cache'
+import { CacheInterface, CacheResponseInterface } from './interfaces/cache'
 
 const Cache = () => {
-    // cache response
+    //  cache data is queried and set by the HeaderForm.
+    //  cache data is used by the CacheTable.
+
+    // cache response state
     const [cacheData, setCacheData] = useState<CacheInterface[]>([])
 
     // display all cache on first page load
