+++ frontend/src/pages/cache/interfaces/cache.ts
